requirements.txt
langchain_community
tiktoken
langchainhub
langchain
langgraph
langserve==0.0.37
pydantic==1.10.12
httpx==0.24.1
python-dotenv
google-generativeai
langchain-google-genai
langchain-groq
langchain-nomic
langchain-text-splitters
sentence-transformers 
streamlit
langchain-openai
ipykernel
chromadb 
fastapi==0.95.2
uvicorn==0.22.0